requirements.txt
Pillow
Flask
BytesIO